Original ESXimager Code from 2012
Written by John Pucila and I (Matt Tentilucci) for a class project at
RIT in 2012. ESXimager in this state utilized netcat to transfer dd
images directly across the network from an ESXi server to an imaging
machine. This process of using netcat sometimes yielded a good image but
sometimes did not for unknown reasons. Interaction with the program was
done entirely through the command line. The code remained in this state
for some time, and was posted on SourceForge
(https://sourceforge.net/projects/esxiimaging/).