Reviving ESXimager Project
In 2014 I decided to pick this project back up and improve upon the
original idea for my thesis at Penn State. Utilizing the Original 2012
code as a baseline, ESXimager was changed to utilize SFTP to transfer an
image from an ESXi server to an imaging machine, instead of netcat. DD
outputs to the ESXi server and the image is then transferred to the
imaging via SFTP. Exact notes written of changes:
Switching to utilize Net::SFTP::Foreign to transfer images across the
network instead of piping DD to netcat as this was causing problems
image is dd to a local storage on ESXi server then copied across the
network to the acquiring computer via SFTP