GUI Improvements, Added Configuration File
Redesign user selection of VMs window and switched from grid to pack
geometry manager. Instead of having a sub window, there will be a frame
within main window that will be updated with the VM choices for the user
to image. This should be a much cleaner look and prevent multiple
windows from popping up. Also added configuration file and
Tools->Settings menu bar for editing it