GUI Improvements, Cleanup after imaging is complete
Improved VM imaging window. Asks the user what VM they want to image,
then what files from that VM, then confirms selection. Added windows
telling the user what the program is doing, when the file gets DD, or
hashes are being calculated because the program will not respond to user
inputs when those things are being executed. Controlled where subwindows
are shown on the screen, they show up in the middle of the main window.
Changed console log box to scrolled and tied STDOUT to print in the box.
Now using print will print in the consoleLog, also
$consoleLog->see('end') shows the bottom of the console log and
essentially makes it scroll automatically as it grows. After the image
has been dd'd and SFTP'd the script will cleanup the .dd files is
created on the ESXi server. Made the checkbutton frame scrolled when
user has to select what vms/files they want to image